+++ NEWS.md
@@ -1,6 +1,9 @@
 ## Version 0.3
 
 * Incorporated Expert-Feedback and updated the protocol with additional fields and explanations.
+* Added glossary, to be further expanded in future versions #4
+* Added tooltip to the protocol for each fields (can be toggled off) #6
+* Added a bookmark option to store the current setting of the protocol as server id (experimental)
 
 ## Version 0.2